Separate HardNormalModuleFactoryPlugin
